Fixes a bug where you can't change the handle of a page once the pages translation is deleted

+++ lib/class.TManager.php
@@ -462,9 +462,12 @@
 			else{
 				foreach( $changed as $translation ){
 					/* @var $translation Translation */
-					$translation->delete();
-					$translation->setHandle($old_handle);
-					$translation->setName($old_handle);
+					if(is_object($translation))
+					{
+						$translation->delete();
+						$translation->setHandle($old_handle);
+						$translation->setName($old_handle);
+					}
 				}
 			}
